Issue #171: refactor build_dataset
