+++ test/output/test_config.py
@@ -2,7 +2,7 @@
 Unit test for special config settings
 """
 
-from deepreg.util import train_and_predict_with_config
+from deepreg.test_util import train_and_predict_with_config
 
 
 def test_unpaired_unlabeled():
