+++ test/output/test_grouped_labeled.py
@@ -1,7 +1,7 @@
 """
 Unit test for grouped labeled data
 """
-from deepreg.util import train_and_predict_with_config
+from deepreg.test_util import train_and_predict_with_config
 
 
 def test_grouped_labeled():
