+++ test/output/test_networks.py
@@ -2,7 +2,7 @@
 Unit test for special config settings
 """
 
-from deepreg.util import train_and_predict_with_config
+from deepreg.test_util import train_and_predict_with_config
 
 
 def test_dvf_network():
