+++ test/output/test_unpaired_unlabeled.py
@@ -2,7 +2,7 @@
 Unit test for unpaired unlabeled data
 """
 
-from deepreg.util import train_and_predict_with_config
+from deepreg.test_util import train_and_predict_with_config
 
 
 def test_unpaired_unlabeled():
