+++ test/unit/test_train.py
@@ -10,8 +10,7 @@ import os
 import pytest
 import tensorflow as tf
 
-from deepreg.dataset.loader.interface import DataLoader
-from deepreg.train import build_callbacks, build_config, build_dataset, train
+from deepreg.train import build_callbacks, build_config, train
 
 
 def test_build_config():
@@ -38,39 +37,6 @@ def test_build_config():
     # checkpoint path ends with h5
     with pytest.raises(ValueError):
         build_config(config_path=config_path, log_dir=log_dir, ckpt_path="example.h5")
-
-
-def test_build_dataset():
-    """
-    Test build_dataset by checking the output types
-    """
-
-    # init arguments
-    config_path = "deepreg/config/unpaired_labeled_ddf.yaml"
-    log_dir = "test_build_dataset"
-    ckpt_path = ""
-
-    # load config
-    config, log_dir = build_config(
-        config_path=config_path, log_dir=log_dir, ckpt_path=ckpt_path
-    )
-
-    # build dataset
-    data_out_train, data_out_val = build_dataset(
-        dataset_config=config["dataset"],
-        preprocess_config=config["train"]["preprocess"],
-    )
-
-    data_loader_train, dataset_train, steps_per_epoch_train = data_out_train
-    data_loader_val, dataset_val, steps_per_epoch_val = data_out_val
-
-    # check output types
-    assert isinstance(data_loader_train, DataLoader)
-    assert isinstance(dataset_train, tf.data.Dataset)
-    assert isinstance(steps_per_epoch_train, int)
-    assert isinstance(data_loader_val, DataLoader)
-    assert isinstance(dataset_val, tf.data.Dataset)
-    assert isinstance(steps_per_epoch_val, int)
 
 
 def test_build_callbacks():
